backend-multidimensional-audio-index/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
audioread==3.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dnspython==2.4.2
email-validator==2.1.0.post1
faiss-cpu==1.7.4
fastapi==0.104.1
fastdtw==0.3.4
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
msgpack==1.0.7
numba==0.58.1
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pandas==2.1.3
platformdirs==4.0.0
pooch==1.8.0
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
Rtree==1.1.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
